+++ src/Sulu/Bundle/AdminBundle/Resources/public/js/components/header/main.js
@@ -373,8 +373,8 @@ define([], function () {
         startLanguageChanger: function() {
             if (!!this.options.toolbarLanguageChanger) {
                 var $element = this.sandbox.dom.createElement(this.sandbox.util.template(templates.languageChanger)({
-                    title: this.options.toolbarLanguageChanger.preSelected || this.sandbox.sulu.user.locale
-                })),
+                        title: this.options.toolbarLanguageChanger.preSelected || this.sandbox.sulu.user.locale
+                    })),
                     options = constants.languageChangerDefaults;
                 this.sandbox.dom.show(this.$find(constants.rightSelector));
                 this.sandbox.dom.append(this.$find(constants.rightSelector), $element);